Added some multithreading functionality to load chunks via web workers. Can be enabled by setting MULTITHREADED to true in World.js
Modified choosing cubes with right faces such that cubes adjecent to a chunk boundary will have the face that is touching the boundary be empty - this is a quick hack for speed boost, but introduces some artifacts.

The main bottleneck is the geometry construction of each chunk on the webgl tier. This takes a long time, and right now, I don't foresee any way to offload it to a web worker without heavily modifying the source code for THREE.js.

The reason why multithreading seems to suffer in performance is that geometry construction is interspersed throughout time, instead of all done after all chunks are loaded. Chunks are loaded when a player crosses a chunk boundary, and in single-thread version, all new chunks are loaded and constructed. In the multi-threaded version, chunks are loaded in parallel, and when a chunk is loaded, geometry is constructed on the main thread. So if you want consistent average frames per second, use the threaded version. If you want pauses each time you cross a chunk boundary, use the single threaded version.